middleware: new docs
Also update the docs based on the new runtime
