+++ rlog/rlog.go
@@ -4,41 +4,10 @@
 // For more information about logging inside Encore applications see https://encore.dev/docs/observability/logging.
 package rlog
 
-// Debug logs a debug-level message.
-// The variadic key-value pairs are treated as they are in With.
-func Debug(msg string, keysAndValues ...interface{}) {
-	panic("encore apps must be run using the encore command")
-}
-
-// Info logs an info-level message.
-// The variadic key-value pairs are treated as they are in With.
-func Info(msg string, keysAndValues ...interface{}) {
-	panic("encore apps must be run using the encore command")
-}
-
-// Error logs an error-level message.
-// The variadic key-value pairs are treated as they are in With.
-func Error(msg string, keysAndValues ...interface{}) {
-	panic("encore apps must be run using the encore command")
-}
-
 // Ctx holds additional logging context for use with the Infoc and family
 // of logging functions.
 type Ctx struct {
 	_ int // for godoc to show unexported fields
-}
-
-// With adds a variadic number of fields to the logging context.
-// The keysAndValues must be pairs of string keys and arbitrary data.
-func With(keysAndValues ...interface{}) Ctx {
-	panic("encore apps must be run using the encore command")
-}
-
-// With creates a new logging context that inherits the context
-// from the original ctx and adds additional context on top.
-// The original ctx is not affected.
-func (Ctx) With(keysAndValues ...interface{}) Ctx {
-	panic("encore apps must be run using the encore command")
 }
 
 // Debug logs a debug-level message, merging the context from ctx
@@ -59,5 +28,12 @@ func (Ctx) Info(msg string, keysAndValues ...interface{}) {
 // with the additional context provided as key-value pairs.
 // The variadic key-value pairs are treated as they are in With.
 func (Ctx) Error(msg string, keysAndValues ...interface{}) {
+	panic("encore apps must be run using the encore command")
+}
+
+// With creates a new logging context that inherits the context
+// from the original ctx and adds additional context on top.
+// The original ctx is not affected.
+func (Ctx) With(keysAndValues ...interface{}) Ctx {
 	panic("encore apps must be run using the encore command")
 }
