+++ storage/sqldb/sqldb.go
@@ -8,8 +8,7 @@ import (
 	"database/sql"
 )
 
-// An error satisfying ErrNoRows is reported by Scan
-// when QueryRow doesn't return a row.
+// ErrNoRows is an error reported by Scan when QueryRow doesn't return a row.
 // It must be tested against with errors.Is.
 var ErrNoRows = sql.ErrNoRows
 
@@ -20,29 +19,6 @@ type ExecResult interface {
 	RowsAffected() int64
 }
 
-// Exec executes a query without returning any rows.
-// The args are for any placeholder parameters in the query.
-//
-// See (*database/sql.DB).ExecContext() for additional documentation.
-func Exec(ctx context.Context, query string, args ...interface{}) (ExecResult, error) {
-	panic("encore apps must be run using the encore command")
-}
-
-// Query executes a query that returns rows, typically a SELECT.
-// The args are for any placeholder parameters in the query.
-//
-// See (*database/sql.DB).QueryContext() for additional documentation.
-func Query(ctx context.Context, query string, args ...interface{}) (*Rows, error) {
-	panic("encore apps must be run using the encore command")
-}
-
-// QueryRow executes a query that is expected to return at most one row.
-//
-// See (*database/sql.DB).QueryRowContext() for additional documentation.
-func QueryRow(ctx context.Context, query string, args ...interface{}) *Row {
-	panic("encore apps must be run using the encore command")
-}
-
 // Tx is a handle to a database transaction.
 //
 // See *database/sql.Tx for additional documentation.
@@ -50,85 +26,25 @@ type Tx struct {
 	_ int // for godoc to show unexported fields
 }
 
-// Begin opens a new database transaction.
-//
-// See (*database/sql.DB).Begin() for additional documentation.
-func Begin(ctx context.Context) (*Tx, error) {
-	panic("encore apps must be run using the encore command")
-}
-
 // Commit commits the given transaction.
 //
 // See (*database/sql.Tx).Commit() for additional documentation.
-func (*Tx) Commit() error {
-	panic("encore apps must be run using the encore command")
-}
+func (*Tx) Commit() error { panic("encore apps must be run using the encore command") }
 
 // Rollback rolls back the given transaction.
 //
 // See (*database/sql.Tx).Rollback() for additional documentation.
-func (*Tx) Rollback() error {
-	panic("encore apps must be run using the encore command")
-}
+func (*Tx) Rollback() error { panic("encore apps must be run using the encore command") }
 
-// Exec is like Exec but executes the query in the given transaction.
-//
-// See (*database/sql.Tx).ExecContext() for additional documentation.
 func (*Tx) Exec(ctx context.Context, query string, args ...interface{}) (ExecResult, error) {
 	panic("encore apps must be run using the encore command")
 }
 
-// Query is like Query but executes the query in the given transaction.
-//
-// See (*database/sql.Tx).QueryContext() for additional documentation.
 func (*Tx) Query(ctx context.Context, query string, args ...interface{}) (*Rows, error) {
 	panic("encore apps must be run using the encore command")
 }
 
-// QueryRow is like QueryRow but executes the query in the given transaction.
-//
-// See (*database/sql.Tx).QueryRowContext() for additional documentation.
 func (*Tx) QueryRow(ctx context.Context, query string, args ...interface{}) *Row {
-	panic("encore apps must be run using the encore command")
-}
-
-// Commit commits the given transaction.
-//
-// See (*database/sql.Tx).Commit() for additional documentation.
-// Deprecated: use tx.Commit() instead.
-func Commit(tx *Tx) error {
-	panic("encore apps must be run using the encore command")
-}
-
-// Rollback rolls back the given transaction.
-//
-// See (*database/sql.Tx).Rollback() for additional documentation.
-// Deprecated: use tx.Rollback() instead.
-func Rollback(tx *Tx) error {
-	panic("encore apps must be run using the encore command")
-}
-
-// ExecTx is like Exec but executes the query in the given transaction.
-//
-// See (*database/sql.Tx).ExecContext() for additional documentation.
-// Deprecated: use tx.Exec() instead.
-func ExecTx(tx *Tx, ctx context.Context, query string, args ...interface{}) (ExecResult, error) {
-	panic("encore apps must be run using the encore command")
-}
-
-// QueryTx is like Query but executes the query in the given transaction.
-//
-// See (*database/sql.Tx).QueryContext() for additional documentation.
-// Deprecated: use tx.Query() instead.
-func QueryTx(tx *Tx, ctx context.Context, query string, args ...interface{}) (*Rows, error) {
-	panic("encore apps must be run using the encore command")
-}
-
-// QueryRowTx is like QueryRow but executes the query in the given transaction.
-//
-// See (*database/sql.Tx).QueryRowContext() for additional documentation.
-// Deprecated: use tx.QueryRow() instead.
-func QueryRowTx(tx *Tx, ctx context.Context, query string, args ...interface{}) *Row {
 	panic("encore apps must be run using the encore command")
 }
 
@@ -186,42 +102,5 @@ func (*Row) Scan(dest ...interface{}) error {
 }
 
 func (*Row) Err() error {
-	panic("encore apps must be run using the encore command")
-}
-
-// constStr is a string that can only be provided as a constant.
-//
-type constStr string
-
-// Named returns a database object connected to the database with the given name.
-//
-// The name must be a string literal constant, to facilitate static analysis.
-func Named(name constStr) *Database {
-	panic("encore apps must be run using the encore command")
-}
-
-type Database struct {
-	_ int // for godoc to show unexported fields
-}
-
-func (*Database) Exec(ctx context.Context, query string, args ...interface{}) (ExecResult, error) {
-	panic("encore apps must be run using the encore command")
-}
-
-func (*Database) Query(ctx context.Context, query string, args ...interface{}) (*Rows, error) {
-	panic("encore apps must be run using the encore command")
-}
-
-func (*Database) QueryRow(ctx context.Context, query string, args ...interface{}) *Row {
-	panic("encore apps must be run using the encore command")
-}
-
-func (*Database) Begin(ctx context.Context) (*Tx, error) {
-	panic("encore apps must be run using the encore command")
-}
-
-// Stdlib returns a *sql.DB object that is connected to the same db,
-// for use with libraries that expect a *sql.DB.
-func (*Database) Stdlib() *sql.DB {
 	panic("encore apps must be run using the encore command")
 }
